fix console warnings
